# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Self-hosted webmail application API backed by IMAP and SMTP.",
        "title": "Go-Cubemail API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/compose/draft": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Saves an email draft (currently a stub; returns ok immediately).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "compose"
                ],
                "summary": "Save draft",
                "responses": {
                    "200": {
                        "description": "status ok",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/compose/send": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Composes and delivers an email via SMTP, then saves a copy to the Sent folder.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "compose"
                ],
                "summary": "Send email",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sender email (defaults to session user)",
                        "name": "from_email",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Sender display name",
                        "name": "from_name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated recipient list",
                        "name": "to",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated CC list",
                        "name": "cc",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated BCC list",
                        "name": "bcc",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Email subject",
                        "name": "subject",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "HTML body",
                        "name": "body_html",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Plain-text body",
                        "name": "body_plain",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Reply-To address",
                        "name": "reply_to",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "File attachments",
                        "name": "attachments",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status sent",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "502": {
                        "description": "SMTP delivery error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/compose/upload": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Pre-uploads an attachment for use in a composed email (currently a stub).",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "compose"
                ],
                "summary": "Upload attachment",
                "responses": {
                    "200": {
                        "description": "files list (empty)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "CookieAuth": {
            "description": "The session cookie 'gorc_session' is required for authenticated endpoints.",
            "type": "apiKey",
            "name": "Cookie",
            "in": "header"
        }
    }
}